Change what's displayed in "kotlinc -version"
In CLI and kotlin-maven-plugin, display the version of JRE the compiler
is running on, do not display the "info:" prefix.

Note that this makes JPS plugin not display the version of the compiler
in the Messages view, since it's no longer parsed as an "info" compiler
message. This is fine, because the JPS plugin reports its own version
itself, which is equal to the compiler version anyway

 #KT-6536 Fixed